Implementación de autenticación parte #2 y proteccion de rutas

+++ routes/api.php
@@ -24,27 +24,27 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 Route::prefix('products')->group(function () {
     Route::get('listPrices', [ProductController::class, 'getPrices']);
-    Route::post('', [ProductController::class, 'store']);
-    Route::put('/{id}', [ProductController::class, 'update']);
+    Route::post('', [ProductController::class, 'store'])->middleware('auth:api');
+    Route::put('/{id}', [ProductController::class, 'update'])->middleware('auth:api');
     Route::get('', [ProductController::class, 'index']);
-    Route::delete('/{id}', [ProductController::class, 'destroy']);
+    Route::delete('/{id}', [ProductController::class, 'destroy'])->middleware('auth:api');
     Route::get('/{id}', [ProductController::class, 'find']);
 });
 
 Route::prefix('categories')->group(function () {
-    Route::post('', [CategoriesController::class, 'store']);
-    Route::put('/{id}', [CategoriesController::class, 'update']);
+    Route::post('', [CategoriesController::class, 'store'])->middleware('auth:api');
+    Route::put('/{id}', [CategoriesController::class, 'update'])->middleware('auth:api');
     Route::get('', [CategoriesController::class, 'index']);
-    Route::delete('/{id}', [CategoriesController::class, 'destroy']);
+    Route::delete('/{id}', [CategoriesController::class, 'destroy'])->middleware('auth:api');
     Route::get('/{id}', [CategoriesController::class, 'find']);
 });
 
 Route::prefix('users')->group(function () {
     Route::post('', [UserController::class, 'store']);
-    Route::get('', [UserController::class, 'index']);
+    Route::get('', [UserController::class, 'index'])->middleware('auth:api');
     Route::get('/{id}', [UserController::class, 'find']);
-    Route::put('/{id}', [UserController::class, 'update']);
-    Route::delete('/{id}', [UserController::class, 'destroy']);
+    Route::put('/{id}', [UserController::class, 'update'])->middleware('auth:api');
+    Route::delete('/{id}', [UserController::class, 'destroy'])->middleware('auth:api');
 });
 
 Route::prefix('auth')->group(function () {
